Change output file names
This PR:
- deleted old `ecCodes` based `nc2grib.py` and rename `grib2io.py` to
`nc2grib.py`
- renamed output file names suggested in issue #38 

Closes issue #38 

TODO:
- update `ensstat` and `TC_tracker` to take new file names
