+++ environment.yml
@@ -7,10 +7,6 @@ dependencies:
   - netcdf4
   - xarray
   - cartopy
-  - pygrib
-  - eccodes
-  - scitools-iris
-  - iris-grib
   - jupyterlab
   - boto3
   - pip
