+++ oper/run_graphcast.py
@@ -27,6 +27,7 @@ from graphcast import graphcast
 from graphcast import normalization
 from graphcast import rollout
 
+from utils.nc2grib import Netcdf2Grib
 
 class GraphCastModel:
     def __init__(
@@ -38,15 +39,13 @@ class GraphCastModel:
         output_dir = None, 
         num_pressure_levels = 13, 
         forecast_length = 64,
-        grb2method = "iris",
     ):
         self.pretrained_model_path = pretrained_model_path
         self.gdas_data_path = gdas_data_path
         self.forecast_length = forecast_length
-        self.num_pressure_levels = num_pressure_levels
         self.case_name = case_name
+        self.num_pressure_levels = num_pressure_levels
         self.config_file_path = config_file
-        self.grb2method = grb2method
         
         if output_dir is None:
             self.output_dir = os.getcwd()
@@ -207,27 +206,13 @@ class GraphCastModel:
         ds = ds.isel(time=slice(1, 2))
         ds['time'] = ds['time'] - pd.Timedelta(hours=6)
 
-        if self.grb2method == "iris":
-            from utils.nc2grib import Netcdf2Grib
 
-            converter = Netcdf2Grib()
-            converter.save_grib2(self.dates[0][1], ds, self.case_name, self.output_dir)
+        converter = Netcdf2Grib(self.dates[0][1], case_name=self.case_name)
+        converter.save_grib2(ds, self.output_dir)
 
-            # Call and save forecasts in grib2
-            converter.save_grib2(self.dates[0][1], forecasts, self.case_name, self.output_dir)
+        # Call and save forecasts in grib2
+        converter.save_grib2(forecasts, self.output_dir)
 
-        elif self.grb2method == "grib2io":
-            from utils.grib2io import Netcdf2Grib
-
-            converter = Netcdf2Grib(self.dates[0][1], case_name=self.case_name)
-            converter.save_grib2(ds, self.output_dir)
-
-            # Call and save forecasts in grib2
-            converter.save_grib2(forecasts, self.output_dir)
-
-        else:
-            raise ValueError(f"Method {self.method} is not supported. Choose either 'iris' or 'grib2io'!")
-        
     def upload_to_s3(self, keep_data):
         s3 = boto3.client('s3')
         
@@ -290,12 +275,11 @@ if __name__ == "__main__":
     parser.add_argument("-c", "--config", help="GC weight member file", default=None)
     parser.add_argument("-o", "--output", help="output directory", default=None)
     parser.add_argument("-p", "--pressure", help="number of pressure levels", default=13)
-    parser.add_argument("-m", "--grb2method", help="method to convert netCDF files to grib2 files", default="iris")
     parser.add_argument("-u", "--upload", help="upload input data as well as forecasts to noaa s3 bucket (yes or no)", default = "no")
     parser.add_argument("-k", "--keep", help="keep input and output after uploading to noaa s3 bucket (yes or no)", default = "no")
     
     args = parser.parse_args()
-    runner = GraphCastModel(args.weights, args.input, args.case_name, args.config, args.output, int(args.pressure), int(args.length), args.grb2method)
+    runner = GraphCastModel(args.weights, args.input, args.case_name, args.config, args.output, int(args.pressure), int(args.length))
     
     runner.load_pretrained_model()
     runner.load_gdas_data()
